Pipelines/requirements.txt
tweepy==4.12.1
python-decouple==3.7